Add quadrotor test and example
