+++ examples/Project.toml
@@ -1,5 +1,6 @@
 [deps]
 COSMOAccelerators = "bbd8fffe-5ad0-4d78-a55e-85575421b4ac"
+Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 CoordinateTransformations = "150eb455-5306-5404-9cee-2592286d6298"
 GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
 Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
