--- examples/visualization/Dropped Text.txt
@@ -1,1 +1,0 @@
-uploads_files_2493288_CGAXR_MC_Drone_High_OBJ
