+++ test/runtests.jl
@@ -31,8 +31,3 @@ end
     include("se3_kinematics_test.jl")
     include("se3_force_test.jl")
 end
-
-
-# @testset "SE3 Examples" begin
-#     include("se3_test.jl")
-# end
